TEC-4142: Improve WizardItem Docs separating Label, Point and Root stories
